% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{CML}
\alias{CML}
\title{A function for getting the sets C, M, L from an ADMG given a treatment}
\usage{
CML(graph, treatment)
}
\arguments{
\item{graph}{A graph object generated via function \code{make.graph()}.}

\item{treatment}{A character string indicating the treatment variable.}
}
\value{
A list of sets C, M, L
}
\description{
Set C is the pre-treatment variable
Set L is the post-treatment variables that are within the district of the treatment, including the treatment itself.
Set M is the rest of the variables
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
CML(graph, treatment='A')

}
\keyword{CML}
