% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{count.districts}
\alias{count.districts}
\title{Get the number of districts in a graph.}
\usage{
\method{count}{districts}(graph)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}
}
\value{
A list with the following two components:
\describe{
\item{districts}{A list of vectors, each vector contains the vertices in a district.}
\item{n.districts}{The number of districts in the graph.}}
}
\description{
Function to count the number of districts in a graph object.
Number of districts in a graph is the number of bidirected connected components in the graph.
A vertex that is not bidirected connected to any other vertex is also considered as a district by convention.
For example, a graph like A B <-> C has two districts: {A, B} and {C}.
}
\keyword{ADMG}
\keyword{district}
\keyword{graph}
