% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{data_backdoor}
\alias{data_backdoor}
\title{This is data generated from a classic backdoor model, where the pre-treatment variable X cause A and Y, and the treatment A cause Y.}
\format{
A data frame with 2000 rows and 3 variables: X, A, Y.
\describe{
\item{X}{a variable that follows uniform distribution at the interval of \eqn{[0,1]}}
\item{A}{a binary treatment}
\item{Y}{a variable that follows normal distribution}
}
}
\usage{
data_backdoor
}
\description{
This is data generated from a classic backdoor model, where the pre-treatment variable X cause A and Y, and the treatment A cause Y.
}
\references{
\url{https://github.com/annaguo-bios/ADMGtmle/tree/main}
}
\keyword{data}
