% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{data_example_a}
\alias{data_example_a}
\title{Data generated from the simulation model in Figure 4(a) of the paper \url{https://arxiv.org/abs/2409.03962}.
The model can also be found in Figure (a) of the github repository: \url{https://github.com/annaguo-bios/flexCausal}}
\format{
A data frame with 2000 rows and 7 variables: X, A, U, M=(M.1,M.2), L, Y.
\describe{
\item{X}{a variable that follows uniform distribution at the interval of \eqn{[0,1]}}
\item{A}{a binary treatment}
\item{U}{an unmeasured confounder following a normal distribution}
\item{M.1}{first component of the bivariate mediator M, following a normal distribution}
\item{M.2}{second component of the bivariate mediator M, following a normal distribution}
\item{L}{a variable that follows normal distribution}
\item{Y}{a variable that follows normal distribution}
}
}
\usage{
data_example_a
}
\description{
Data generated from the simulation model in Figure 4(a) of the paper \url{https://arxiv.org/abs/2409.03962}.
The model can also be found in Figure (a) of the github repository: \url{https://github.com/annaguo-bios/flexCausal}
}
\references{
\url{https://github.com/annaguo-bios/flexCausal}
}
\keyword{data}
