% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{data_frontdoor}
\alias{data_frontdoor}
\title{Data generated from a classic front-door model.}
\format{
A data frame with 2000 rows and 5 variables:
\describe{
\item{X}{a pre-treatment variable following a uniform distribution on \eqn{[0, 1]}}
\item{U}{an unmeasured confounder following a normal distribution}
\item{A}{a binary treatment variable (0/1)}
\item{M}{a binary mediator variable (0/1)}
\item{Y}{a continuous outcome variable following a normal distribution}
}
}
\usage{
data_frontdoor
}
\description{
Data generated from a front-door model where a pre-treatment variable X and
an unmeasured confounder U affect both the treatment A and outcome Y, and
the treatment A affects Y through a binary mediator M.
}
\references{
\url{https://github.com/annaguo-bios/flexCausal}
}
\keyword{data}
