% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.adj_matrix}
\alias{f.adj_matrix}
\title{Get adjacency matrix from graph.}
\usage{
f.adj_matrix(graph)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}
}
\value{
An adjacency matrix of the graph.
}
\description{
Function to perform topological sorting using Kahn's algorithm. The treatment variable is ranked with the largest order possible if there are ties.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.adj_matrix(graph)
}
\keyword{ADMG}
\keyword{graph}
\keyword{ordering}
