% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.children}
\alias{f.children}
\title{Get the children of a node OR nodes in a graph.}
\usage{
f.children(graph, nodes)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{nodes}{A character vector of nodes for which to extract children.}
}
\value{
A vector of vertices contains children set of the given nodes.
}
\description{
Function to extract the children of a node OR nodes in a graph object.
Children of a node are the nodes that have edges from the given node.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.children(graph, c('A'))

}
\keyword{ADMG}
\keyword{children}
\keyword{graph}
