% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.district}
\alias{f.district}
\title{Get the district of a vertex in a graph.}
\usage{
f.district(graph, node)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{node}{A character string of a vertex for which to extract district.}
}
\value{
A vector of vertices that is in the district of the given vertex.
}
\description{
Function to extract the name of vertices that is in the district of a given vertex in a graph object.
District of a unfixed vertex Vi is the set of vertices that are connected to Vi by bidirected edges, including Vi itself by convention.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.district(graph, c('A'))

}
\keyword{ADMG}
\keyword{district}
\keyword{graph}
