% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.markov_blanket}
\alias{f.markov_blanket}
\title{Get the Markov blanket of a vertex in a graph.}
\usage{
f.markov_blanket(graph, node)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{node}{A character string of a vertex for which to extract Markov blanket.}
}
\value{
A vector of vertices that is in the Markov blanket of the given vertex.
}
\description{
Function to get the Markov blanket of a vertex in a graph object.
Markov blanket of a vertex Vi is the union of vertices that is in the district of Vi and their parents set.
Mb= union(district(Vi), parents(district(Vi))). THIS IS PSUEDO CODE.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.markov_blanket(graph, 'A')

}
\keyword{ADMG}
\keyword{blanket}
\keyword{graph}
\keyword{markov}
