% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.markov_pillow}
\alias{f.markov_pillow}
\title{Get the Markov pillow of a vertex in a graph.}
\usage{
f.markov_pillow(graph, node, treatment = NULL)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{node}{A character string of a vertex for which to extract Markov pillow.}

\item{treatment}{A character string specifying the treatment variable in the graph.}
}
\value{
A vector of vertices that is in the Markov pillow of the given vertex.
}
\description{
Function to get the Markov pillow of a vertex in a graph object.
Markov pillow of a vertex Vi is the subset of the Markov blanket of Vi that proceed Vi in the topological ordering of the graph.
Mp= {Vj in union(district(Vi), parents(district(Vi))): Vj proceed Vi}. THIS IS PSUEDO CODE.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.markov_pillow(graph, 'A')

}
\keyword{ADMG}
\keyword{graph}
\keyword{markov}
\keyword{pillow}
