% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.parents}
\alias{f.parents}
\title{Get the parents of a node OR nodes in a graph.}
\usage{
f.parents(graph, nodes)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{nodes}{A character vector of nodes for which to extract parents.}
}
\value{
A vector of vertices contains parents set of the given nodes.
}
\description{
Function to extract the parents of a node OR nodes in a graph object.
Parents of a node are the nodes that have directed edges pointing to the node.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.parents(graph, c('Y','L'))

}
\keyword{ADMG}
\keyword{graph}
\keyword{parents}
