% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.reachable_closure}
\alias{f.reachable_closure}
\title{Reachable closure of a set of vertices in a graph.}
\usage{
f.reachable_closure(graph, nodes)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{nodes}{A character vector of vertices.}
}
\value{
A list containing the following components:
\describe{
\item{reachable_closure}{A character vector containing the reachable closure of the given vertices.}
\item{fixing_order}{A character vector of vertices telling the order in which the vertices were fixed.}
\item{graph}{The CADMG obtained via recursively fixing as many vertices as possible in the set of all vertices (V) excluding the set of vertices specified by the \code{nodes} parameter (S), i.e. V \ S. }}
}
\description{
Function to return the reachable closure of a set of vertices in a graph object.
First obtain a Conditional ADMG (CADMG) via recursively fixing as many vertices as possible in the set of all vertices (V) excluding the set of vertices specified by the \code{nodes} parameter (S), i.e. V \ S.
The reachable closure is the subset of V \ S, where each vertex is not fixable even upon fixing other vertices.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.reachable_closure(graph, 'A')

}
\keyword{ADMG}
\keyword{closure}
\keyword{graph}
\keyword{reachable}
