% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.top_order}
\alias{f.top_order}
\title{Get the topological ordering of a graph from graph object.}
\usage{
f.top_order(graph, treatment = NULL)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{treatment}{A character string indicating the treatment variable.
If NULL, this function will rank vertices according to their input order in the vertices vector when there are ties.}
}
\value{
A vector of vertices ranked with rank from small to large.
}
\description{
Function to extract the adjacency matrix from a graph object.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
f.top_order(graph)

}
\keyword{ADMG}
\keyword{adjancency.matrix}
\keyword{graph}
