% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{f.top_orderMAT}
\alias{f.top_orderMAT}
\title{Topological ordering of a graph from adjacency matrix.}
\usage{
f.top_orderMAT(adj.matrix, treatment = NULL)
}
\arguments{
\item{adj.matrix}{An adjacency matrix of the graph.}

\item{treatment}{A character string indicating the treatment variable.}
}
\value{
A vector of vertices ranked with rank from small to large.
}
\description{
Function to perform topological sorting using Kahn's algorithm. The treatment variable is ranked with the largest order possible if there are ties.
}
\keyword{ADMG}
\keyword{graph}
\keyword{ordering}
