% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{fix}
\alias{fix}
\title{Perform fixing operation on the given vertices in a graph.}
\usage{
fix(graph, nodes)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{nodes}{A chacater vector of vertices to fix in the graph.}
}
\value{
A graph object with the given vertices fixed.
}
\description{
Function to perform the fixing operation on the given vertices in a graph object.
Fixing a vertex in a graph means deleting all the incoming directed edges to the vertex and deleting all the bidirected edges concerning this vertex. The vertex is also marked as fixed.
This function is called by the \code{f.reachable_closure()} function.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
fix(graph, 'A')

}
\keyword{ADMG}
\keyword{fix}
\keyword{graph}
