% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{insert}
\alias{insert}
\title{Replace element at a specific position.}
\usage{
insert(vec, pos, value)
}
\arguments{
\item{vec}{A vector.}

\item{pos}{An integer indicating the position to replace element in vec with value.}

\item{value}{A vector to insert into vec.}
}
\value{
A vector with value inserted at position pos.
}
\description{
This function replace element at a specific position of a vector called vec with another vector called value.
}
\examples{
insert(c('A','B','C'), 2, 'D')
}
\keyword{element}
\keyword{replace}
\keyword{vector}
