% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{is.fix}
\alias{is.fix}
\title{Fixability of a treatment variable in a graph.}
\usage{
is.fix(graph, treatment)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}

\item{treatment}{A character string specifying the treatment variable in the graph.}
}
\value{
A logical value indicating whether the treatment is primal fixable.
}
\description{
Function to check if a treatment variable is fixable in a graph object.
If the treatment is fixable, then the average causal effect of the treatment on any choice of the outcome in the given graph is always identified via backdoor adjustment.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
is.p.fix(graph, 'A')

}
\keyword{ADMG}
\keyword{fixable}
\keyword{graph}
\keyword{primal}
