% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{is.mb.shielded}
\alias{is.mb.shielded}
\title{Check if a graph is mb-shielded.}
\usage{
is.mb.shielded(graph)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}
}
\value{
A logical value indicating whether the graph is mb-shielded.
}
\description{
Function to check if a graph is mb-shielded.
A graph being mb-shielded means that the graph only implies ordinary equality constraints on the observed data distribution.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
is.mb.shielded(graph)
}
\keyword{ADMG}
\keyword{graph}
\keyword{mb-shielded}
