% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{is.np.saturated}
\alias{is.np.saturated}
\title{Check if a graph is nonparametrically saturated.}
\usage{
is.np.saturated(graph)
}
\arguments{
\item{graph}{A graph object generated by the \code{make.graph()} function.}
}
\value{
A logical value indicating whether the graph is nonparametrically saturated.
}
\description{
Function to check if a graph is nonparametrically saturated.
A graph being nonparametrically saturated means that the graph implies NO equality constraints on the observed data distribution
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))
is.np.saturated(graph)

}
\keyword{ADMG}
\keyword{graph}
\keyword{nonparametrically}
\keyword{saturated}
