% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{make.graph}
\alias{make.graph}
\title{Create graph object.}
\usage{
make.graph(vertices, bi_edges, di_edges, multivariate.variables = NULL)
}
\arguments{
\item{vertices}{A character vector of vertices in the graph.}

\item{bi_edges}{A list of vectors that record the bidirectional edges in the graph. For example, \code{bi_edges=list(c('A','B'))} means there is a bidirectional edge between vertex A and B.}

\item{di_edges}{A list of vectors that record the directed edges in the graph. For example, \code{di_edges=list(c('A','B'))} means there is a directed edge from vertex A to B.}

\item{multivariate.variables}{A list of variables that are multivariate in the causal graph.
For example, \verb{multivariate.variables=list(M=c('M1,'M2'))} means M is bivariate and the corresponding columns in the dataframe are M1 and M2. Default is \code{NULL}.}
}
\value{
A graph object with the following components:
\describe{
\item{\code{vertices}}{Equivalent to the input argument \code{vertices}.}
\item{\code{fixed}}{A data frame with a column \code{fixed} that indicates whether the vertex is fixed or not. The vertices is not fixed initially.}
\item{\code{bi_edges}}{Equivalent to the input argument \code{bi_edges}.}
\item{\code{di_edges}}{Equivalent to the input argument \code{di_edges}.}
\item{\code{multivariate.variables}}{A list of variables that are multivariate in the causal graph. For example, \verb{multivariate.variables=list(M=c('M1,'M2'))} means M is bivariate and the corresponding columns in the dataframe are M1 and M2. Default is \code{NULL}.}}
}
\description{
This function create a graph object that can be used in other functions in this package.
}
\examples{
make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')))

}
\keyword{ADMG}
\keyword{graph}
