% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{rerank}
\alias{rerank}
\title{Rerank the target vector according to the variable order in reference vector}
\usage{
rerank(target, reference)
}
\arguments{
\item{target}{A vector to be re-ranked}

\item{reference}{A vector that provides the order for re-ranking}
}
\value{
A re-ranked vector
}
\description{
In our case, we re-rank the vector C, M, L according to topological order tau
}
\examples{
rerank(target=c('A','C','B'), reference=c('A','B','C','D'))

}
\keyword{ranking}
\keyword{rerank}
