% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_density_ratio_dnorm.R
\name{calculate_density_ratio_dnorm}
\alias{calculate_density_ratio_dnorm}
\title{Calculate the density ratio of M given its Markov pillow at treatment level a0 verse a1.}
\usage{
calculate_density_ratio_dnorm(a0, M, graph, treatment, data, formula = NULL)
}
\arguments{
\item{a0}{A numeric value indicating the treatment level. The density ratio is calculated at this level verse level 1-a0.}

\item{M}{A character string indicating the name of the variable, for which the density ratio is calculated.}

\item{graph}{A graph object created by \code{\link{make.graph}}.}

\item{treatment}{A character string indicating the name of the treatment variable A.}

\item{data}{A data frame containing the data.}

\item{formula}{A regression formula specifying the relationship between M or L and its Markov pillow. See the description for \code{ADMGtmle()} for more details.}
}
\value{
A vector of density ratios for each row of the data.
}
\description{
This function calculates the density ratio of two conditional densities. Let \eqn{mp(M)-A} denote the Markov pillow of M excluding A. The density ratio is calculated as \eqn{p(M|mp(M)-A, A=a0)/p(M|mp(M)-A, A=a1)}.
If M is univariate or multivariate continuous variable, it assume \eqn{p(M|mp(M)} is normally distributed. If M is a binary variable, it assume \eqn{p(M|mp(M)} is Bernoulli distributed.
}
\examples{
graph <- make.graph(vertices=c('A','M','L','Y','X'),
bi_edges=list(c('A','Y')),
di_edges=list(c('X','A'), c('X','M'), c('X','L'),
c('X','Y'), c('M','Y'), c('A','M'), c('A','L'), c('M','L'), c('L','Y')),
multivariate.variables = list(M=c('M.1','M.2')))
ratio <- calculate_density_ratio_dnorm(a0=0, "M", graph, treatment="A", data=data_fig_4a)
head(ratio)

}
\keyword{density}
\keyword{graph}
\keyword{ratio}
