% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{data_example_a}
\alias{data_example_a}
\title{This is data generated from the simulation model in Figure 4(a) of the paper "Targeted Learning of Average Causal Effects in Graphical Models with Unmeasured Variables: Extensions of the Front-Door Model".
The model can also be found in Figure (a) of the github repository: \url{https://github.com/annaguo-bios/ADMGtmle/tree/main}}
\format{
A data frame with 2000 rows and 6 variables: X, A, M, L, Y.
\describe{
\item{X}{a variable that follows uniform distribution at the interval of \eqn{[0,1]}}
\item{A}{a binary treatment}
\item{M}{a variable that contains M.1 and M.2, generated from a bivariate normal distribution}
\item{L}{a variable that follows normal distribution}
\item{Y}{a variable that follows normal distribution}
}
}
\usage{
data_example_a
}
\description{
This is data generated from the simulation model in Figure 4(a) of the paper "Targeted Learning of Average Causal Effects in Graphical Models with Unmeasured Variables: Extensions of the Front-Door Model".
The model can also be found in Figure (a) of the github repository: \url{https://github.com/annaguo-bios/ADMGtmle/tree/main}
}
\references{
\url{https://github.com/annaguo-bios/ADMGtmle/tree/main}
}
\keyword{data}
